Solution to Toll Plaza Management System

+++ Toll Plaza Management System@@67877d61dc107486ad5c2750/code.c
@@ -4,14 +4,14 @@
 struct Vehicle{
     char number[20];
     char type[10];
-    float toll;
+    int toll;
 };
 
 int main(){
     int n;
     scanf("%d",&n);
     struct Vehicle  arr[n];
-    float arr1[3]={0.0,0.0,0.0};
+    int arr1[3]={0,0,0};
 
     for(int i=0;i<n;i++){
         scanf("%s %s %d",&arr[i].number,&arr[i].type,&arr[i].toll);
